MAAP-AWS-Arcee/loader/requirements.txt
python-multipart==0.0.12
fastapi[standard]==0.115.2
humanize==4.11.0
langchain-community==0.3.2
pypdf==5.0.1
boto3==1.35.40
langchain-mongodb==0.2.0
pymongo==4.10.1
langchain-aws==0.2.2
unstructured[all-docs]==0.15.14
langchain-unstructured[all-docs]==0.1.5
python-dotenv==1.0.1
